intake-mcp/mcp-server/requirements.txt
fastmcp
pyyaml
python-dotenv
httpx
starlette